+++ packages/python/m/cli/commands/ci/celt.py
@@ -5,7 +5,8 @@ from m.cli import command, validate_json_payload, validate_payload
 from pydantic import BaseModel, Field
 
 from ....core import one_of
-from ...utils import display_issue
+
+# from ...utils import display_issue
 
 
 class Arguments(BaseModel):
