+++ packages/python/m/core/__init__.py
@@ -1,10 +1,11 @@
 # noqa: WPS412
 from .fp import Bad, Good, OneOf
-from .issue import Issue
+from .issue import Issue, IssueDict
 from .one_of import issue, non_issue, one_of
 
 __all__ = [  # noqa: WPS410
     'Issue',
+    'IssueDict',
     'issue',
     'non_issue',
     'one_of',
